Item1805: Disabled checking if Validation Method is set to 'none'. Also fixed debugging message to avoid a warning.
git-svn-id: http://svn.foswiki.org/trunk@4495 0b4bb1d4-4e5a-0410-9cc4-b2b747904278

+++ core/lib/Foswiki/UI.pm
@@ -516,6 +516,9 @@ See Foswiki::Validation for more information.
 sub checkValidationKey {
     my ($session) = @_;
 
+    # If validation is disabled, do nothing
+    return if ( $Foswiki::cfg{Validation}{Method} eq 'none' );
+
     # Check the nonce before we do anything else
     my $nonce = $session->{request}->param('validation_key');
     $session->{request}->delete('validation_key');
